refactor: move 5 leaf packages to internal/muxcore/
Move ipc, jsonrpc, remap, serverid, classify from internal/ to
internal/muxcore/ as Phase 1 of muxcore library extraction.

All 5 packages have zero internal dependencies — pure leaf packages.
16 consumer files updated with new import paths.

No logic changes. All 10 test packages pass.
