+++ internal/daemon/daemon.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/thebtf/mcp-mux/internal/control"
 	"github.com/thebtf/mcp-mux/internal/mux"
-	"github.com/thebtf/mcp-mux/internal/serverid"
+	"github.com/thebtf/mcp-mux/internal/muxcore/serverid"
 	"github.com/thejerf/suture/v4"
 )
