+++ internal/daemon/daemon_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/thebtf/mcp-mux/internal/control"
-	"github.com/thebtf/mcp-mux/internal/ipc"
+	"github.com/thebtf/mcp-mux/internal/muxcore/ipc"
 )
 
 func testLogger(t *testing.T) *log.Logger {
