+++ internal/daemon/snapshot_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/thebtf/mcp-mux/internal/classify"
+	"github.com/thebtf/mcp-mux/internal/muxcore/classify"
 	"github.com/thebtf/mcp-mux/internal/control"
 	"github.com/thebtf/mcp-mux/internal/mux"
 )
