+++ internal/mux/client.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"sync"
 
-	"github.com/thebtf/mcp-mux/internal/ipc"
+	"github.com/thebtf/mcp-mux/internal/muxcore/ipc"
 )
 
 // RunClient connects to an existing owner via IPC and bridges
