+++ internal/muxcore/ipc/ipc_test.go
