+++ internal/muxcore/jsonrpc/extra_test.go
