+++ internal/muxcore/jsonrpc/scanner.go
